# api/openapi.yaml
openapi: 3.1.0
info:
  title: Seva API
  description: |
    API specification for the Seva platform.
    This is the single source of truth that generates Go server stubs,
    SvelteKit TypeScript client, and Flutter Dart client.
  version: 1.0.0
  contact:
    name: Platform Engineering
    email: engineering@seva.io
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.staging.seva.io
    description: Staging
  - url: https://api.seva.io
    description: Production
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/jobs/{id}/review:
    post:
      operationId: submitReview
      summary: Submit review for completed job
      description: Submits a review for a completed job. Both customer and provider can submit reviews for each other.
      tags: [Jobs]
      parameters:
        - $ref: "#/components/parameters/JobIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewCreate"
            example:
              rating: 5
              comment: Excellent work! Fixed the leak quickly and cleaned up after.
      responses:
        "201":
          description: Review submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
              example:
                id: "rev_01HS2NKFP3R7E3DYGZM2PFWT4C"
                job_id: "job_01HRZ3KFWBN9T1E7YGZM2PFWT4"
                reviewer_id: "usr_01HQXK5V8R3N0E7YGZM2PFWT4A"
                reviewee_id: "usr_01HQXK5V8R3N0E7YGZM2PFWT4B"
                rating: 5
                comment: Excellent work! Fixed the leak quickly and cleaned up after.
                created_at: "2026-03-17T15:00:00Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # ---------------------------------------------------------------------------
  # Categories
  # ---------------------------------------------------------------------------
...
  /api/v1/reviews:
    get:
      operationId: listReviews
      summary: List reviews for current user
      description: Lists reviews written by or about the current user.
      tags: [Reviews]
      parameters:
        - name: role
          in: query
          description: Filter by role in the review (reviewer or reviewee)
          schema:
            type: string
            enum: [reviewer, reviewee]
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/LimitQuery"
      responses:
        "200":
          description: Reviews retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Review"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/reviews/{id}/respond:
    post:
      operationId: respondToReview
      summary: Respond to a review
      description: Allows a provider to respond to a review left by a customer.
      tags: [Reviews]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [response]
              properties:
                response:
                  type: string
                  maxLength: 1000
      responses:
        "200":
          description: Response recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/providers/{id}/reviews:
    get:
      operationId: listProviderReviews
      summary: List reviews for a provider
      tags: [Reviews]
      security: []
      parameters:
        - $ref: "#/components/parameters/ProviderIdPath"
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/LimitQuery"
      responses:
        "200":
          description: Provider reviews retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Review"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/providers/{id}/ratings:
    get:
      operationId: getProviderRatingStats
      summary: Get provider rating statistics
      tags: [Reviews]
      security: []
      parameters:
        - $ref: "#/components/parameters/ProviderIdPath"
      responses:
        "200":
          description: Rating statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RatingStats"
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------------------------------------------------------------------------
  # Notifications
  # ---------------------------------------------------------------------------
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from the /auth/verify-otp or /auth/refresh endpoints.

  # ---------------------------------------------------------------------------
  # Reusable Parameters
  # ---------------------------------------------------------------------------
  parameters:
...
    ProviderIdPath:
      name: id
      in: path
      required: true
      description: Provider (user) ID
      schema:
        type: string
      example: "usr_01HQXK5V8R3N0E7YGZM2PFWT4A"
    JobIdPath:
      name: id
      in: path
      required: true
      description: Job ID
      schema:
        type: string
      example: "job_01HRZ3KFWBN9T1E7YGZM2PFWT4"
...
    PageQuery:
      name: page
      in: query
      description: Page number (1-indexed)
      schema:
        type: integer
        minimum: 1
        default: 1
      example: 1
    LimitQuery:
      name: limit
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      example: 20

  # ---------------------------------------------------------------------------
  # Reusable Responses
  # ---------------------------------------------------------------------------
  responses:
    BadRequest:
      description: Bad request — invalid input or missing required fields
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: BAD_REQUEST
            message: Invalid request parameters
            details:
              - field: phone
                message: Phone number must include country code
    Unauthorized:
      description: Unauthorized — missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: UNAUTHORIZED
            message: Authentication required
            details: null
    Forbidden:
      description: Forbidden — insufficient permissions for the requested action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: FORBIDDEN
            message: You do not have permission to perform this action
            details: null
    NotFound:
      description: Not found — the requested resource does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: NOT_FOUND
            message: Resource not found
            details: null
    UnprocessableEntity:
      description: Unprocessable entity — the request is well-formed but semantically invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: UNPROCESSABLE_ENTITY
            message: Validation failed
            details:
              - field: rating
                message: Rating must be between 1 and 5
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: INTERNAL_SERVER_ERROR
            message: An unexpected error occurred
            details: null

  # ---------------------------------------------------------------------------
  # Schemas
  # ---------------------------------------------------------------------------
  schemas:
...
    ReviewCreate:
      type: object
      required: [rating, comment]
      properties:
        rating:
          type: integer
          description: Rating from 1 to 5
          minimum: 1
          maximum: 5
        comment:
          type: string
          description: Review text
          maxLength: 2000
    Review:
      type: object
      required: [id, job_id, reviewer_id, reviewee_id, rating, comment, created_at]
      properties:
        id:
          type: string
          description: Unique review identifier
          examples:
            - "rev_01HS2NKFP3R7E3DYGZM2PFWT4C"
        job_id:
          type: string
          description: Job this review is for
        reviewer_id:
          type: string
          description: User who wrote the review
        reviewee_id:
          type: string
          description: User being reviewed
        rating:
          type: integer
          minimum: 1
          maximum: 5
          description: Rating (1-5 stars)
        comment:
          type: string
          description: Review text
          maxLength: 2000
        created_at:
          type: string
          format: date-time
          description: Review creation timestamp

    # --- Disputes ---
...
    PaginatedResponse:
      type: object
      required: [data, total, page, limit, has_more]
      properties:
        data:
          type: array
          items: {}
          description: Array of result items
        total:
          type: integer
          description: Total number of items matching the query
          examples:
            - 42
        page:
          type: integer
          description: Current page number (1-indexed)
          examples:
            - 1
        limit:
          type: integer
          description: Number of items per page
          examples:
            - 20
        has_more:
          type: boolean
          description: Whether more pages are available
          examples:
            - true
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Machine-readable error code
          examples:
            - "BAD_REQUEST"
        message:
          type: string
          description: Human-readable error message
          examples:
            - "Invalid request parameters"
        details:
          description: Additional error details (validation errors, etc.)
          oneOf:
            - type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
            - type: "null"

    # --- Notifications ---
...
    RatingStats:
      type: object
      properties:
        avg_rating:
          type: number
          format: double
        total_reviews:
          type: integer
        distribution:
          type: object
          properties:
            "1":
              type: integer
            "2":
              type: integer
            "3":
              type: integer
            "4":
              type: integer
            "5":
              type: integer

    # --- KYC ---
...